redirected models to the new models folder
