+++ client/src/app/project/[projectid]/scheduleComponents/CalendarColumnTimeHeader.tsx
@@ -1,5 +1,3 @@
-import { FC } from "react";
-
 const CalendarColumnTimeHeader = () => {
   return (
     <div className="w-16 pr-1">
