+++ client/src/app/project/[projectid]/scheduleComponents/CalendarColumns.tsx
@@ -3,7 +3,7 @@ import {
   EachScheduleData,
   ProjectData,
   ScheduleHeaderData,
-} from "@/util/models";
+} from "@/util/models/ProjectModels";
 import EachScheduleItem from "./EachScheduleItem";
 
 interface Props {
